fix: torch version issue avail 1.9.1

+++ networks/scheduler/ddpm.py
@@ -40,7 +40,7 @@ class DDPMSampler:
         variance = (1 - alpha_prod_t_prev) / (1 - alpha_prod_t) * current_beta_t
 
         # we always take the log of variance, so clamp it to ensure it's not 0
-        variance = torch.clamp(variance, min=1e-20)
+        variance = torch.clamp(variance.to(torch.float32), min=1e-20)
 
         return variance
